# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  mydb.StructNewsClass:
    properties:
      admin_id:
        description: 管理员id
        type: integer
      create_time:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: id
        type: integer
      is_hot:
        description: 是否热门
        type: boolean
      is_recommend:
        description: 是否推荐
        type: boolean
      is_show:
        description: 是否显示
        type: boolean
      keywords:
        description: 关键词
        type: string
      name:
        description: 新闻分类名称
        type: string
      parent_id:
        description: 父级分类id
        type: integer
      sort:
        description: 排序
        type: integer
      status:
        description: 状态:0=禁用,1=启用
        type: integer
      update_time:
        description: 更新时间
        type: integer
    type: object
...
  util.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: nav.fandoc.org
info:
  contact: {}
  description: This is a sample server for a nav web site.
  title: Nav Web Site API
  version: "1.0"
paths:
...
  /news/addClass:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 添加新的新闻分类
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 新闻分类名称
        in: formData
        name: name
        required: true
        type: string
      - description: 父级分类id
        in: formData
        name: parent_id
        type: integer
      - description: 排序
        in: formData
        name: sort
        type: integer
      - description: 图标
        in: formData
        name: icon
        type: string
      - description: 描述
        in: formData
        name: description
        type: string
      - description: 关键词
        in: formData
        name: keywords
        type: string
      - description: 是否显示
        in: formData
        name: is_show
        type: boolean
      - description: 是否推荐
        in: formData
        name: is_recommend
        type: boolean
      - description: 是否热门
        in: formData
        name: is_hot
        type: boolean
      - description: 状态:0=禁用,1=启用
        in: formData
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 新闻分类添加成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 添加新闻分类失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 添加新闻分类
      tags:
      - news
...
  /news/deleteClass/{id}:
    delete:
      description: 根据新闻分类ID删除新闻分类
      parameters:
      - description: 新闻分类ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 新闻分类删除成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 删除新闻分类失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 删除新闻分类
      tags:
      - news
...
  /news/getClassDetail/{id}:
    get:
      description: 根据新闻分类ID获取新闻分类详情
      parameters:
      - description: 新闻分类ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取新闻分类详情成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/mydb.StructNewsClass'
                message:
                  type: string
              type: object
        "500":
          description: 获取新闻分类详情失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取新闻分类详情
      tags:
      - news
  /news/getClassList:
    get:
      description: 获取所有新闻分类的列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取新闻分类列表成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/mydb.StructNewsClass'
                  type: array
                message:
                  type: string
              type: object
        "500":
          description: 获取新闻分类列表失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取新闻分类列表
      tags:
      - news
...
  /news/updateClass/{id}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: 根据新闻分类ID修改新闻分类信息
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 新闻分类ID
        in: path
        name: id
        required: true
        type: string
      - description: 新闻分类名称
        in: formData
        name: name
        required: true
        type: string
      - description: 父级分类id
        in: formData
        name: parent_id
        type: integer
      - description: 排序
        in: formData
        name: sort
        type: integer
      - description: 图标
        in: formData
        name: icon
        type: string
      - description: 描述
        in: formData
        name: description
        type: string
      - description: 关键词
        in: formData
        name: keywords
        type: string
      - description: 是否显示
        in: formData
        name: is_show
        type: boolean
      - description: 是否推荐
        in: formData
        name: is_recommend
        type: boolean
      - description: 是否热门
        in: formData
        name: is_hot
        type: boolean
      - description: 状态:0=禁用,1=启用
        in: formData
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 新闻分类修改成功
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: 修改新闻分类失败
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 修改新闻分类
      tags:
      - news
...
swagger: "2.0"